Minor update to script.js and index.html
Made changes to the message that the user receives at the end of the quiz, rather than a generic "congratulations" message, I have given the user feedback so that they can see their exact score. I removed the <p> element from index.html as it was now redundant and replaced with innerHTML js.