Minor update to script.js
Changed the message that users are given when completing the quiz. Users now are given one of two messages depending on their score. I have also added console.log to log the correct/incorrect answer in the console and the score the user currently has, this is for testing purposes (removed during implementation by mistake).